pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contract-graph"
version = "0.1.0"
description = "Cross-boundary contract intelligence for heterogeneous codebases"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "sauremilk" }]
keywords = ["contracts", "type-sync", "static-analysis", "cross-boundary", "pydantic", "typescript"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
typescript = ["tree-sitter>=0.22", "tree-sitter-typescript>=0.22"]
sarif = ["sarif-om>=1.0"]
dev = ["pytest>=8.0", "pytest-cov", "ruff", "pyright"]
all = ["contract-graph[typescript,sarif,dev]"]

[project.scripts]
contract-graph = "contract_graph.cli:main"

[project.urls]
Homepage = "https://github.com/sauremilk/contract-graph"
Repository = "https://github.com/sauremilk/contract-graph"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
